Introduce a segmentation API.
This change introduces an API for manipulating a segment of an ordered
collection. This is a bit like a slice, but does not require the
collection to be contiguous and, most importantly, supports topological
mutations (i.e., insertions and removals). Rather than implement various
takes on clearing, removing, splitting, etc. for `NonEmpty` collection
types, these kinds of errors are instead a part of segmentation: it is
an error to isolate a segment of a `NonEmpty` collection that is not a
strict subset.

This implementation features sound drain iterators that maintain the
non-empty invariant if leaked. This change also bumps the MSRV to
`1.76.0`.