Refactor `Remainder` and `Iterator1::maybe_empty` functions.
This change renames the `Remainder` type defintion to `Result` and
restricts its usage to contexts wherein an empty iterator is an error
condition, namely `Iterator1::try_from_iter`. Elsewhere, remainders are
represented as a bare `Iterator`, leaving any conversion to `Iterator1`
up to client code.

To better support `Iterator1::maybe_empty` functions, which query the
iterator, this change also introduces `Query` and the `Matched` and
`IsMatched` type definitions. These allow both the output and remainder
of a query to be used ergonomically in certain situations where first
converting into an `Iterator` with a `let` binding is more cumbersome.