Implement `Saturate` more explicitly for collections.
This change replaces the blanket implementation of `Saturate` bound on
`Extend` and `Vacancy` with more explicit implementations for
collections. This avoids an incorrect implementation for relational
collection types with some notion of capacity (that implement `Extend`
and `Vacancy`). In particular, `BinaryHeap` may implement these traits
and but requires a different `Saturate` implementation.