Introduce an unsafe ordering trait for non-empty B-trees.
This change introduces the `Ord1` trait. This trait is unsafe and has a
bound on `Ord`. `Ord1` must maintain a consistent total ordering for
implementing types per the documentation for `BTreeSet` and `BTreeMap`.
It is unsafe, because failing to maintain this invariant is UB in
certain non-empty relational collection APIs, such as segmentation for
non-empty B-trees and `BTreeSet1::split_off_tail`. These operations rely
on a consistent total ordering to guarantee that collections remain
non-empty and so are unsound without a bound on `Ord1`.

Without `Ord1`, safe code can cause a segment to remove items beyond a
strict subset and drain a non-empty collection, for example. This is
difficult to achieve, but possible! A similar problem exists for any
potential API that removes items within a relational range based on an
`Ord` implementation.

This change also introduces `pop_{first,last}_until_only` APIs, which
can remove all items but the last without requiring `Ord1`. To implement
`Ord1` for networking types migrated to `core`, the MSRV has been bumped
to `1.77.0`.