Rename `saturate` to `saturate_and` for consistency.
This change renames `saturate` and `saturated` functions to
`saturate_and` and `saturated_and`, respectively. The "and" suffix is
used when advancing an iterator to produce a `Feed` (both an output and
a remainder). `saturate` and `saturated` functions remain, but discard
any remainder and only produce an output collection instead. This aligns
with other extensions, like `QueryAnd`.