Remove `Iterator1::collect`.
This change removes `collect` in favor of only `collect1` in
`Iterator1`. The distinction between maybe-empty and non-empty
collections in `FromIterator` implementations is required, because of a
spurious coherence error. While this distinction is necessary,
`Iterator1::collect` should instead be reached via
`Iterator1::into_iter().collect()` as with other similar functions that
exhibit no different behavior for non-empty iterators.