Remove bespoke `const` functions from `NonEmpty`.
It can be useful to construct `Slice1` in `const` contexts, but
`FromMaybeEmpty::from_maybe_empty_unchecked` cannot be used for this,
because `const` functions are not yet supported in traits. To support
`const` construction, `NonEmpty` had bespoke `const` functions. This
change removes those functions in favor of inline calls to
`mem::transmute` in `Slice1`, which is less noisy