Rename `TakeOr::only` functions to `TakeOr::get_only`.
The `only` functions of the `TakeOr` proxy use standard Rust naming
conventions locally to the type, but in context this name is a bit
unusual (and incongruent with other `TakeOr` functions). Since `TakeOr`
is a proxy type and its methods follow "or" functions, the "get" is more
appropriate. For example, `xs.pop_or().get_only()` is similarly named to
the conventional non-proxy form `xs.pop_or_get_only()`.