Remove unsound `splice` functions.
This change removes `splice` functions from non-empty `Vec` types,
because these functions are unsound!

Like `Drain`, the `Splice` iterator defensively establishes a consistent
state when constructed and then restores the desired state when dropped.
(In fact, `Splice` contains a `Drain` iterator.) This means that leaking
the iterator prevents restoration to the desired state. For prefix
ranges, this leaves a `Vec` empty regardless of the arguments to
`splice`. It is not enough for the spliced items to originate from an
`Iterator1`: the items are never inserted if the `Splice` iterator is
leaked.

`Splice` iterators can deploy the swap technique used by
`vec1::SwapDrainSegment`. However, more is required for segments,
because the ultimate range of the segment cannot be known until the
iterator is dropped (unlike `Drain`). Moreover, `splice` must assert
that range arguments are a strict subset if implemented for `Vec1`,
defying the role of segmentation.

Though `splice` functions are very efficient in very particular
circumstances, they are mostly a convenience, so there is limited
utility in supported them for non-empty collections.