Return an `Error` type from fallible constructions.
This change introduces `EmptyError`, which is returned in the `Err`
variant of `Result`s in fallible conversions and constructions for
non-empty types.

`EmptyError` contains the empty collection used in the conversion.
Unlike returning the collection directly, `EmptyError` implements the
standard `Error` trait. This better supports error propagation and
reporting while continuing to avoid dropping the input collection as
before.

`ArrayVec1` is unique in that some of its conversions have two failure
modes: the input is empty or exceeds the capacity of the array. These
operations now return a `CardinalityError`, which composes `EmptyError`
and `CapacityError`.