Introduce a trait for isomorphic ordering.
This change introduced the `UnsafeIsomorph` trait. Implementations
promise, at risk of memory safety, that `Self` and the input type
represent the same data with equivalent ordering. For example, `Vec` and
slices have this property. This trait provides a mechanism to use such
isomorphic types in APIs where this contract is necessary for memory
safety.