[wip] Use an `enum` for relational ranges.
This change begins to refactor `RelationalRange` into an `enum` with two
variants: `Item` and `Trim`. The `Item` variant contains an `ItemRange`,
which is what `RelationalRange` was previously. This allows segments
over relational collections to have only one type using only one range,
but at the cost of branching on that range.