Lift expressions out of `unsafe` blocks in `vec1` macro.
This change lifts the expressions used to construct `Vec`s in the `vec1`
macro out of `unsafe` blocks. This prevents the unannotated use of
unsafe code directly in `vec1` expression metavariables. This example
builds successfully before this change:

```rust
let x = 0u8;
let x = &x as *const u8;

let xs = vec1![*x];
```

The pointer dereference `*x` is unsafe, but `vec1!` expands the
expression within an unseen `unsafe` block. After this change, this is
no longer the case, and so `unsafe { *x }` is now required in this
example.