Rename take APIs with a `_if_many` suffix.
This change renames APIs that take items out of non-empty collections.
The `Take` type is now named `TakeIfMany` and its various type
definitions have been renamed similarly. Collection APIs now have this
same suffix, so `Vec1::pop` is now `Vec1::pop_if_many`.

This change is somewhat verbose, but makes the distinction and behavior
of these functions much more clear. In the expression
`xs.pop().or_none()`, the `or_none` may raise an eyebrow, but it is not
clear that it concerns `pop`'s behavior. After this change, this
expression is now `xs.pop_if_many().or_none()`, which at least
communicates that `pop_if_many` is **not** the same as `Vec::pop`.