Implement `as_mut_{inner}` conversions for non-empty collections.
This change introduces unsafe functions for getting a mutable reference
to the underlying representation of non-empty collections. For example,
`Vec1::as_mut_vec` returns a mutable reference to its underlying `Vec`.

Calling code must guarantee that collections are non-empty when the
mutable reference is dropped (including downgrades to shared references)
and the `NonEmpty` can be accessed again. These conversions should be
avoided, but are necessary for certain in-place operations that may be
significantly more expensive to implement via moves and into/from
conversions.