Provide `retain_until_only` for non-empty collections.
This change implements `retain_until_only` for some non-empty
collections, including `Vec1` and `BTreeMap1`. Unlike `try_retain`,
these functions operate over a mutable reference and retain an
unspecified range of the collection that omits a single item. This item
is then compared against the predicate and, if it were to be removed, a
reference to the item is returned (and it remains in the collection).