Simplify and rename segmentation traits.
This change removes a segmentation traits and renames the remaining
traits.

Previously, the `SegmentedBy` and `Segmentation` traits worked together,
with `SegmentedBy` providing the primary implementation and
`Segmentation` lowering the input type parameters to the function-level.
However, this is rarely necessary. When type inference fails, the syntax
can be unfortunate, but this does not occur often. What was previously
called `Segmentation` has been removed.

Here is a summary of the changed/renamed traits:

- `Segmentation` => **removed**
- `SegmentedBy` => `Query`
- `SegmentedOver` => `Segmentation`