Introduce `UnsafeHash` and `Except`.
This change introduces the `hash` module, `UnsafeHash` trait, and the
`Except` type constructor for `HashSet1`.

`UnsafeHash` is analogous to `UnsafeOrd` and requires compliance in `Eq`
and `Hash` implementations for memory safety. This is needed for
`HashSet1::except`, because `Except` relies on the presence and ability
to consistently identify its `key` to maintain the non-empty guarantee.

`Except` is somewhat analgous to `Segment`, but for unordered
collections and a bit more limited. Importantly, it can be used to clear
all but one item in a `HashSet1`.