Implement de/serialization for unsized and reference types.
This change improves the integration with `serde`. Before this change,
de/serialization was implemented through conversions into and from a
`Serde` proxy type. This did not support some `Serialize` and
`Deserialize` implementations for non-empty types that are available for
their maybe-empty counterparts, such as `Serialize` for `[_]` and
`Deserialize` for `Box<[_]>`.

Now these traits are implemented directly for `NonEmpty<_>` and containers
like `Box<NonEmpty<_>>`. This change also introduces tests for some of
these types.